Attemp to fix murmur starting but not listening on first run
Murmur might not find the correct ip address to start listening on the
initial run. This leaves the service on a funny state, where it is not
listening on any port, but it remains alive.

Documenting the workarounds:

1- Restart the docker container

Running a `docker restart <id>` will kill and start the process again,
while mantaining the network, so mumble starts again and it can find the
appropriate interafaces. This is less then ideal.

2- Specify the ANY interface for IPv6/IPv4

We can set the host option on the ini configuration to listen on any
interface manually, using `::` for IPv6/IPv4, and `0.0.0.0` for IPv4
only. This commit goes this route and it is configured to allow IPv6
connections, even tho the Docker network is not so nice about it yet.

Bug report reference:
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=780300
https://github.com/mumble-voip/mumble/issues/1629